requirements.txt
llama-index>=0.9.0
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
llama-index-llms-openai-like
llama-index-embeddings-huggingface